% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignLifeCycle.R
\name{AssignLifeCycle}
\alias{AssignLifeCycle}
\title{Main module function to assign life cycle category to each household}
\usage{
AssignLifeCycle(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{AssignLifeCycle} assigns the life cycle category to each household
based on the numbers of adults in the household, whether children are in the
household, and whether the adults are retired.
}
\details{
This function assigns the life cycle category to each household based on the
numbers of adults in the household, whether children are in the household,
and whether the adults are retired.
}
