% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AssignLifeCycle.R
\docType{data}
\name{AssignLifeCycleSpecifications}
\alias{AssignLifeCycleSpecifications}
\title{Specifications list for AssignLifeCycle module}
\format{A list containing 3 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
AssignLifeCycle.R script.
}
\usage{
AssignLifeCycleSpecifications
}
\description{
A list containing specifications for the AssignLifeCycle module.
}
\keyword{datasets}
