% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{calcHhAgeTypes}
\alias{calcHhAgeTypes}
\title{Calculate proportions of households by household size}
\usage{
calcHhAgeTypes(Inp_ls = PumsHhInp_ls)
}
\arguments{
\item{Inp_ls}{A list containing the specifications for
the estimation data contained in "pums_hh.csv".}
}
\value{
A matrix where the rows are the household types and the columns are
the age categories and the values are the number of persons.
}
\description{
\code{calcHhAgeTypes} creates a matrix of household types and age
probabilities.
}
\details{
This function produces a matrix of probabilities that a person in one of six
age groups is in one of 524 household types where each household type is
determined by the number of persons in each age category.
}

