% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{createGrpHhByAge}
\alias{createGrpHhByAge}
\title{Create group quarters population for an Azone}
\usage{
createGrpHhByAge(GrpPrsn_Ag)
}
\arguments{
\item{GrpPrsn_Ag}{A named vector containing the number of persons in each age
category.}
}
\value{
A list containing 7 components. Each component is a vector where each
  element of a vector corresponds to a simulated household. The components
  are as follows:
Age0to14 - number of persons age 0 to 14 in the household
Age15to19 - number of persons age 15 to 19 in the household
Age20to29 - number of persons age 20 to 29 in the household
Age30to54 - number of persons age 30 to 54 in the household
Age 55to64 - number of persons age 55 to 64 in the household
Age65Plus - number of persons 65 or older in the household
HhSize - total number of persons in the household
}
\description{
\code{createGroupQtrHhByAge} creates the quarters 'households' for an Azone
where each 'household' is a single person in group quarters.
}
\details{
This function creates a set of simulated 'households' living in group
quaters in an Azone. Each group quarters 'household' is a single person in
each of 6 age categories: 0 to 14, 15 to 19, 20 to 29, 30 to 54, 55 to 64,
and 65 plus.
}
