% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{createHhByAge}
\alias{createHhByAge}
\title{Create simulated households for an Azone}
\usage{
createHhByAge(Prsn_Ap, MaxIter = 100, TargetHhSize = NA,
  TargetProp1PerHh = NA)
}
\arguments{
\item{Prsn_Ap}{A named vector containing the number of persons in each age
category.}

\item{MaxIter}{An integer specifying the maximum number of iterations the
algorithm should use to balance and reconcile the population allocation to
household types.}

\item{TargetHhSize}{A double specifying a household size target value or NA
if there is no target.}

\item{TargetProp1PerHh}{A double specifying a target for the proportion of
households that are one-person households or NA if there is no target.}
}
\value{
A list containing 7 components. Each component is a vector where each
  element of a vector corresponds to a simulated household. The components
  are as follows:
Age0to14 - number of persons age 0 to 14 in the household
Age15to19 - number of persons age 15 to 19 in the household
Age20to29 - number of persons age 20 to 29 in the household
Age30to54 - number of persons age 30 to 54 in the household
Age 55to64 - number of persons age 55 to 64 in the household
Age65Plus - number of persons 65 or older in the household
HhSize - total number of persons in the household
}
\description{
\code{createHhByAge} creates a set of simulated households for an Azone that
reasonably represents a population census or forecast of persons in each of 6
age categories.
}
\details{
This function creates a set of simulated households for an Azone that
reasonably represents the population census or forecast of persons in each of
6 age categories: 0 to 14, 15 to 19, 20 to 29, 30 to 54, 55 to 64, and 65
plus. The function uses a matrix of probabilities that a person in each age
group might be present in each of 524 household types. This matrix
(HtProb_HtAp) is estimated by the calcHhAgeTypes function which is described
above. Household types are distinguished by the number of persons in each age
category in the household. The function fits the distribution of households
by type by iteratively applying the probability matrix to the population,
reconciling households allocated by type based on the population assigned,
recomputing the assigned population, calculating the difference between the
assigned population by age and the input population by age, recalculating the
probabilities, and assigning the population difference. This process
continues until the difference between the assigned population and the input
population by age group is less than 0.1%. After the households are
synthesized, the size of each household is calculated.
}
