% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\name{CreateHouseholds}
\alias{CreateHouseholds}
\title{Main module function to create simulated households}
\usage{
CreateHouseholds(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module along with:
LENGTH: A named integer vector having a single named element, "Household",
which identifies the length (number of rows) of the Household table to be
created in the datastore.
SIZE: A named integer vector having two elements. The first element, "Azone",
identifies the size of the longest Azone name. The second element, "HhId",
identifies the size of the longest HhId.
}
\description{
\code{CreateHouseholds} creates a set of simulated households that each have
a unique household ID, an Azone to which it is assigned, household
size (number of people in the household), and numbers of persons in each of
6 age categories.
}
\details{
This function creates a set of simulated households for the model region
where each household is assigned a household size, an Azone, a unique ID, and
numbers of persons in each of 6 age categories. The function calls the
createHhByAge and createGrpHhByAge functions for each Azone to create
simulated households containing persons by age category from a vector of
persons by age category for the Azone. The list of vectors produced by the
Create Households function are to be stored in the "Household" table. Since
this table does not exist, the function calculates a LENGTH value for the
table and returns that as well. The framework uses this information to
initialize the Households table. The function also computes the maximum
numbers of characters in the HhId and Azone datasets and assigns these to a
SIZE vector. This is necessary so that the framework can initialize these
datasets in the datastore. All the results are returned in a list.
}
