% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictIncome.R
\docType{data}
\name{GQIncModel_ls}
\alias{GQIncModel_ls}
\title{Group quarters income model}
\format{A list having the following components:
\describe{
  \item{Type}{a string identifying the type of model ("linear")}
  \item{Formula}{makeModelFormulaString(IncModel_LM)}
  \item{PrepFun}{a function that prepares inputs to be applied in the linear model}
  \item{OutFun}{a function that transforms the result of applying the linear model}
  \item{Summary}{the summary of the linear model estimation results}
  \item{SearchRange}{a two-element vector specifying the range of search values}
}}
\source{
PredictIncome.R script.
}
\usage{
GQIncModel_ls
}
\description{
A list containing the income model equation and other information needed to
implement the group quarters income model.
}
\keyword{datasets}
