% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateHouseholds.R
\docType{data}
\name{HtProb_HtAp}
\alias{HtProb_HtAp}
\title{Household size proportions}
\format{A matrix having 950 rows (for Oregon data) and 6 colums:}
\source{
CreateHouseholds.R script.
}
\usage{
HtProb_HtAp
}
\description{
A dataset containing the proportions of households by household size.
}
\keyword{datasets}
