% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictHousing.R
\name{PredictHousing}
\alias{PredictHousing}
\title{Main module function to predict the housing type for all households.}
\usage{
PredictHousing(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictWorkers} predicts the housing type for all households.
}
\details{
This function predicts the housing choice of each household. It uses the
estimated models of binomial choice model for determining the probability
that the housing choice for each household is single family (SF) vs.
multifamily (MF). The group quarters population is assigned to group quarters
(GQ).
}
