% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictIncome.R
\name{PredictIncome}
\alias{PredictIncome}
\title{Main module function to predict income for regular households and persons
living in noninstitutional group quarters.}
\usage{
PredictIncome(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictIncome} predicts the income of regular households and persons
living in noninstitutional group quarters.
}
\details{
This function predicts the of each household. It uses the estimated models of
regular household income and group quarters income along with inputs for
average per capita income for households and for group quarters.
}
