% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictWorkers.R
\name{PredictWorkers}
\alias{PredictWorkers}
\title{Main module function to predict workers by age for each household}
\usage{
PredictWorkers(L)
}
\arguments{
\item{L}{A list containing the components listed in the Get specifications
for the module.}
}
\value{
A list containing the components specified in the Set
specifications for the module.
}
\description{
\code{PredictWorkers} predicts the number of workers by age group for each
household and tallies the total number of workers for each household.
}
\details{
This function predicts the number of workers for each household. Household
workers are assigned to 5 age groups: 15-19, 20-29, 30-54, 55-64, and 65Plus.
}
