% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictWorkers.R
\docType{data}
\name{PropHhWkr_HtAg}
\alias{PropHhWkr_HtAg}
\title{Worker proportions}
\format{A a matrix of the proportion of persons in age group who are workers
for each household type.:}
\source{
PredictWorkers.R script.
}
\usage{
PropHhWkr_HtAg
}
\description{
A dataset that contains proportion of workers by age group by household
type for the regular household population and proportion of workers by age
group for noninstitutional group quarters population.
}
\keyword{datasets}
