% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictWorkers.R
\docType{data}
\name{PredictWorkersSpecifications}
\alias{PredictWorkersSpecifications}
\title{Specifications list for PredictWorkers module}
\format{A list containing 4 components:
\describe{
 \item{RunBy}{the level of geography that the module is run at}
 \item{Inp}{module inputs to be saved to the datastore}
 \item{Get}{module inputs to be read from the datastore}
 \item{Set}{module outputs to be written to the datastore}
}}
\source{
PredictWorkers.R script.
}
\usage{
PredictWorkersSpecifications
}
\description{
A list containing specifications for the PredictWorkers module.
}
\keyword{datasets}
